requirements/requirements-base.txt
clearml==1.15.0

hydra-core>=1.3.0

iterative-stratification>=0.1.7
kaggle==1.6.12
matplotlib==3.9.2
numpy==1.23.0
omegaconf>=2.3.0
onnx>=1.15.0
onnxruntime>=1.17.1
onnxruntime-gpu>=1.20.0
onnxsim>=0.4.36
pandas==2.0.0
protobuf==5.28.3
pydantic==2.1.1
scikit-learn==1.5.0

tensorboard==2.18.0
timm>=0.9.6
tqdm==4.67.0
